Moved ImageData and support function into it's own dedicated ImageData.h/.cpp
